Updated the CLI-Argument parser
Previously Tunneload used StructOpt to parse and in general Handle the CLI-Configuration, however
there were some features that I would like for this, that just werent supported and also not really
possible with it. Therefore tunneload now switched to my own args parser "argser", which is first and foremost
made to fullfill all the needs for tunneload, right now all the args are more or less the same than
previously, but in future commits this will most likely change as we add more desired features to it and then
restructure some of the args to use them.