Logging and Parse-Fixes
Switched the Logging over to formatting the Output in JSON-Format, as this makes log aggregation easier
and especially makes it easier to filter/parse out any needed Data.

Then also fixed the Errors when parsing Traefik based Middlewares, which has somehow been broken in one
of the recent commits and went unnoticed for a while.