E2E-Tests for Kubernetes
This commit is the first one to add End-To-End/Integration tests for some kubernetes resources. This
is mainly the first test to make sure the tests work as intended and can be expanded to test to other
parts of kubernetes as well.

To automatically run these tests, I added a new workflow that will setup a local k3s cluster in the
Github-Action and then creates all the needed resources for the tests and then runs the tests against
this newly created and controlled kubernetes cluster without any other external dependencies.