Improved E2E-Tests and cleaned up
First I improved the E2E-Tests by writing a small utility function to compare to vecs without considering
their Order, as this may change between test runs depending on kubernetes and the order is not important
only that the right Elements are in the Vec itself. This was then also used to fix my latest failing
test.

I then also already cleaned up the Code for the other kubernetes-config parts, to already make it easier
for testing once i get to them with E2E-Tests