Added Middleware support for K8s-Ingress
The Kubernetes Ingress Configurator can now also assign Middlewares to ingress Routes based on the
annotations set on the Ingress-Rule. For this it will look for a annotation with the key
'tunneload-middleware' and the Value should be a comma seperated List of the Middleware names.
These will then be applied to the final Rule as with any other Configurator